Refactor Name for ParentTaskGUI
